Deleted individual filters and property filters. Moved all of them to Filters and PropertyFilters Replaced MetadataEvaluator instance methods per static ones, refactored them to accept filter function not called as filter so. Minimal folder refactoring over AutoForm and AutoFormInternal Refactored ExpressionEvaluator to es6 class
