+++ src/metadata/evaluator/filter/PropertyFilters.js
@@ -1,0 +1,24 @@
+import ExpressionEvaluator from '../expression/ExpressionEvaluator.js';
+
+export default class PropertyFilters {
+
+    /**
+     *
+     * @param propertyName
+     * @param propertyValue
+     * @param model
+     * @returns {*}
+     */
+    static filterProperty(propertyName, propertyValue, model) {
+        if(!model) {
+            throw new Error('model is required');
+        }
+
+        if (typeof(propertyValue) === "function" && propertyName.indexOf('$') != 0) {
+            // do something
+            return ExpressionEvaluator.evaluate(propertyValue, model);
+        }
+
+        return propertyValue;
+    }
+}
